+++ src/main/java/it/gov/pagopa/pu/debtpositions/repository/view/installment/InstallmentPIIViewRepositoryImpl.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.debtpositions.repository.view.installment;
 
-import it.gov.pagopa.pu.debtpositions.dto.generated.InstallmentsSearchFiltersDTO;
+import it.gov.pagopa.pu.debtpositions.dto.InstallmentsSearchFiltersDTO;
 import it.gov.pagopa.pu.debtpositions.dto.generated.PagedInstallmentsView;
 import it.gov.pagopa.pu.debtpositions.mapper.PagedInstallmentsViewMapper;
 import org.springframework.data.domain.Pageable;
